+++ apps/bookbuy/src/app/common/libs/models/book.model.ts
@@ -15,3 +15,11 @@ export interface Book {
     saleInfo: any;
     accessInfo: any;
 }
+
+export interface NavItem {
+    displayName: string;
+    disabled?: boolean;
+    iconName: string;
+    route?: string;
+    tag?: boolean;
+}
