+++ apps/bookbuy/src/app/material.module.ts
@@ -42,6 +42,7 @@ import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { MatTreeModule } from '@angular/material/tree';
 import { OverlayModule } from '@angular/cdk/overlay';
+import { MatFormFieldModule } from '@angular/material/form-field';
 
 @NgModule({
     exports: [
@@ -89,6 +90,7 @@ import { OverlayModule } from '@angular/cdk/overlay';
         OverlayModule,
         PortalModule,
         ScrollingModule,
+        MatFormFieldModule
     ]
 })
 export class BookMaterialModule { }
