+++ apps/bookbuy/src/app/services/book.service.spec.ts
@@ -22,10 +22,11 @@ describe('BookService', () => {
   });
 
   it('TC 2: should hit getSearchedBooks API', () => {
-    spyOn(httpClient, 'post').and.returnValue(of(<HttpResponse<any>>{
+    spyOn(httpClient, 'get').and.returnValue(of(<HttpResponse<any>>{
       body: {}
     }));
-    service.getSearchedBooks({ payload: 'java' })
+    service.getSearchedBooks({ payload: 'java' });
+    expect(httpClient.get).toBeCalled();
   });
 
 });
