+++ apps/bookbuy/src/app/shared/components/book-card/book-card.component.spec.ts
@@ -308,6 +308,7 @@ describe('BookCardComponent', () => {
       }
     }
     component.showDetails(book);
+    expect(component.type).toBe('bookdetail');
   });
 
   it('TC 3: should call showList', () => {
@@ -390,7 +391,7 @@ describe('BookCardComponent', () => {
     expect(component.type).toBe('booklist');
   });
 
-  it('TC 5: should call removeFromCart', () => {
+  it('TC 6: should call removeFromCart', () => {
     let book = {
       "kind": "books#volume",
       "id": "6_AqtAEACAAJ",
