+++ apps/bookbuy/src/app/shared/shared.module.ts
@@ -1,9 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
-import { MatButtonModule } from '@angular/material/button';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
 
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
@@ -13,6 +9,7 @@ import { BookEffects } from '@store/effects';
 import { BookReducer } from '@store/reducer';
 
 import { SharedRoutingModule } from './shared-routing.module';
+import { BookMaterialModule } from '../material.module';
 
 import { BookCardComponent } from '@shared-components/book-card/book-card.component';
 
@@ -22,10 +19,7 @@ import { BookCardComponent } from '@shared-components/book-card/book-card.compon
   imports: [
     NgbModule,
     CommonModule,
-    MatIconModule,
-    MatCardModule,
-    MatButtonModule,
-    MatSnackBarModule,
+    BookMaterialModule,
     SharedRoutingModule,
     EffectsModule.forRoot([BookEffects]),
     StoreModule.forRoot({ book: BookReducer })
