+++ apps/bookbuy/src/app/store/effects.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 
 import { EMPTY } from 'rxjs';
-import { catchError, map, mergeMap } from 'rxjs/operators';
+import { catchError, map, switchMap } from 'rxjs/operators';
 
 import { BookService } from '@services/book.service';
 import { ActionTypes } from './actions';
@@ -17,7 +17,7 @@ export class BookEffects {
 
     loadBooks$ = createEffect(() => {
         return this.actions$.pipe(ofType(ActionTypes.LoadItems),
-            mergeMap((action) =>
+            switchMap((action) =>
                 this.bookService.getSearchedBooks(action).pipe(
                     map(books => {
                         return { type: ActionTypes.LoadSuccess, payload: books };
