+++ apps/bookbuy/src/app/store/reducer.spec.ts
@@ -1,5 +1,4 @@
 import { ActionTypes, AddToCart, AddToCollection, LoadItems, RemoveAllFromCart, RemoveFromCart, StoreSearch } from "./actions";
-import { BookStore } from "./models/book.model";
 import { BookReducer, initialState } from "./reducer";
 
 describe('post reducer', () => {
